requirements.txt
streamlit>=1.33,<2
numpy
pandas
plotly
scipy
ezdxf
numpy-stl

